Fix support for comments
Previous implementation of comments was simply converting all "<!--" and
"-->" to "{/*" and "*/}". This was very naive and caused a lot of edge
cases. This also supported nesting Markdown comments in JSX, which isn't
how regular Markdown works; nesting comments inside regular HTML doesn't
work, so it doesn't make sense that it works here.

This fix simply takes advantage of remark's support for comments, which
takes care of all of the edge cases and doesn't reinvent the wheel.

Fixes #256
Fixes #244