fix custom loader example
This makes the code uglier, unfortunately, but the example provided here is incorrect. Since all the code is indented by 4 spaces, it is treated by the markdown parser as a code block, and you end up seeing the exports in the file inside of a formatted code block. Eliminating the extra indentation fixes this.