Preserve positional information for AST nodes
In order to maintain positional metadata for AST nodes
that are parsed, we need to ensure that each import/export
is eaten in order as a correct subvalue.

Since the remark eat function requires the string to be
consumed in order, with no missing characters, the babel
plugin now returns the start value which is used to sort
the nodes and partition the input string. They're then
combined again and returned as nodes that can be eaten
in order by the tokenizer.

This also comes with the benefit of no longer changing the
input code which occurs with the prior babel generator.