Enable remark plugins that insert JSX and HTML
Changed `html`-to-`jsx` node relabelling to happen before remark
plugins are applied instead of after remark plugins are applied. This,
combined with the parsing of `raw` nodes in `HAST`, allows remark
plugins to insert `jsx` nodes AND `html` nodes.

Previously, all `html` nodes were indiscriminately relabelled as
`jsx`, which should only be true for the user-layer when manually
writing MDX and not when `operating on the MDX ASTs with plugins.