Reorganize docs
This commit reorganizes and rewrites the advanced docs and also adds
a more beginner friendly introduction to MDX in Getting Started.

The need for these changes comes from the fact that MDX is now about
a year old and has grown a lot since.  In that growing, a lot of the
docs changed and were shuffled around.  And lots of new stuff was
added.  Additionally, with the 1.0.0 release around the corner, MDX
is going stable and is ready for use by people that aren’t familiar
with stuff like ASTs and jargon-y terms.

This branch supersedes the `more-docs` branch.