raw nodes should be labeled as element after conversion
`raw` returns the root node type as `root`, which ends up causing
double-processing issues. We swapped it to `jsx` to fix that, which ends up
causing the converted HAST to get picked off before it can be processed into
template strings, etc. The type should be `element` so that downstream
processing happens appropriately.