Use babel standalone package
Though it isn't recommended, MDX often finds itself
being used on the client or other environmentts where
bundler configuration might not be customizable.

This switches to babel standalone to allow for better
support for those scenarios.

Before actually merging this it might be worthwhile to
add some benchmarking to ensure that this doesn't introduce
a performance regression (I'm not sure how standalone
works internally).

---

Closes #704