Handle JSX member expressions
JSX member expressions in MDX weren't being properly
handled. This means that they weren't actually rendered
because the MDX type wasn't passed for the pragma.

So, if your MDX contained something like the following:

  <Foo.Bar />

The rendered output was incorrect.

Now, the MDX type is properly added, the shortcode is constructed
if needed, and is pulled out from the components object in
context.