Removed: This is not the place to explain components
I feel it's a good idea to add a note on what components are, but this
hardly feels like the place as we've used *component* a bunch of times
before the reader gets to this point.

Also, if we explain it, I think we can do better than *reusable things*
as that's the kind of explanation that makes sense when you already know
what it is.