Refactor to clarify how to connect providers
The connection of providers (`@mdx-js/react`, `@mdx-js/preact`, or
`@mdx-js/vue`) doesn’t happen with them, but in the integration
(`@mdx-js/esbuild`, `@mdx-js/loader`, `@mdx-js/node-loader`, or
`@mdx-js/rollup`), or `@mdx-js/mdx` itself.
This adds an explanation of that in § Use and also removes some info
in favor of an (already linked) article that explains it in more
detail.

Closes GH-1837.