Use TypeScript `@import` JSDoc tags
Previously we used `@typedef` tags to simulate type imports. The problem
is that are not really imports. They define and export a new type. We
now use `@import` introduced by TypeScript 5.5 to actually import types.

The pattern using `@typedef` is still used in index files to export
types.

This change also replaces the type casting inside MDX files with
definitions of the `Props` types.

This also replaces usage of the global `JSX` namespace with proper
types. For TypeScript<5.1 the correct return type of React components is
`ReactElement`, for Preact `VNode`. For TypeScript>=5.1, the correct
return type for React components is `ReactNode`, for Preact
`ComponentChildren`.